+++ CHANGELOG.md
@@ -27,6 +27,7 @@
   * Add keys to radio groups at MPPLATypeOptionsOverlayWindowWidget.
   * Created 3500-ui_change_line_type_test.
   * Separating THLSizeOrientationPainter from THLinePainter.
+  * Single action to set both LSize and Orientation values at TH2FileEditElementEditController.
 
 ## 0.2.35 - 2026-02-13 - The [Chiquita Bacana's daughter](https://www.youtube.com/watch?v=QajgvvuFOSg) release
 * New features: